requirements.txt
fastapi
uvicorn
pymongo
motor
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
cachetools
pydantic-settings
geopy